Created functional component
created item.js that lists the properties in a div. Then created itemlist.js that creates objects for the items and uses the Item functional component so that it would pass on the properties and place them in a div.  Creates page.js that uses the component so it would show the list. Lastly, styled item.js and itemlist.js div, as well as header.